pyproject.toml
[project]
name = "triangulate"
version = "0.1.0"
description = "Chronological visualization platform that distinguishes verified facts from unverified claims using multi-agent AI"
requires-python = ">=3.11"
dependencies = [
    # Core
    "fastapi>=0.109.0",
    "pydantic>=2.5.0",
    "typer>=0.9.0",
    # AI/LLM
    "langgraph>=0.0.40",
    "litellm>=1.35.0",
    "openai>=1.10.0",
    # Data
    "feedparser>=6.0.10",
    "httpx>=0.26.0",
    "sqlalchemy>=2.0.25",
    "aiosqlite>=0.19.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.2",
    "rich>=13.7.0",
    "toml>=0.10.2",
    "scikit-learn>=1.3.0",
    "beautifulsoup4>=4.14.3",
    "trafilatura>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[project.scripts]
triangulate = "src.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "black>=26.1.0",
]

[tool.pytest.ini_options]
markers = [
    "integration: tests that exercise real external integrations",
]
